+++ src/sphinx/changelog.rst
@@ -11,6 +11,7 @@ Changes in Version 2.5.0
 -  Removed scalaj.collections dependency
 -  Updated the build file
 -  Added unidoc and updated documentation
+-  Updated MongoDBList to handle immutable params
 
 Changes in Version 2.4.1
 ------------------------